+++ spec/eve_online/esi/corporation_assets_spec.rb
@@ -9,7 +9,7 @@ describe EveOnline::ESI::CorporationAssets do
 
   specify { expect(subject).to be_a(EveOnline::ESI::Base) }
 
-  specify { expect(described_class::API_PATH).to eq("/v4/corporations/%<corporation_id>s/assets/") }
+  specify { expect(described_class::API_PATH).to eq("/v5/corporations/%<corporation_id>s/assets/") }
 
   describe "#initialize" do
     context "without options" do
@@ -97,7 +97,7 @@ describe EveOnline::ESI::CorporationAssets do
 
   describe "#path" do
     specify do
-      expect(subject.path).to eq("/v4/corporations/12345678/assets/")
+      expect(subject.path).to eq("/v5/corporations/12345678/assets/")
     end
   end
 
@@ -109,7 +109,7 @@ describe EveOnline::ESI::CorporationAssets do
 
   describe "#url" do
     specify do
-      expect(subject.url).to eq("https://esi.evetech.net/v4/corporations/12345678/assets/?page=1")
+      expect(subject.url).to eq("https://esi.evetech.net/v5/corporations/12345678/assets/?page=1")
     end
   end
 end